Adding a CacheStorage trait for cache methods.
Since the new campaign cache and the previous user cache will have caching methods that are very similar, extracted them out into a CacheStorage trait so that both cache related classes can include the methods into themselves and DRY up the code.
