+++ app/Repositories/UserRepositoryContract.php
@@ -13,22 +13,6 @@ interface UserRepositoryContract extends RepositoryContract
     public function create($account);
 
     /**
-     * Find the specified resource in the database.
-     *
-     * @param  string  $id  Northstar ID
-     * @return object
-     */
-    // public function find($id);
-
-    /**
-     * Get collection of all users or set of users by ids.
-     *
-     * @param  array $ids Northstar IDs
-     * @return \Illuminate\Support\Collection
-     */
-    // public function getAll(array $ids = []);
-
-    /**
      * Get collection of users by the specified role.
      *
      * @param  string $role
